refactor: migrate to socket.io client (NEEDS TO BE TESTED)
