+++ ogybot.csproj
@@ -14,6 +14,7 @@
     <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="8.0.0" />
     <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="8.0.0" />
     <PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.20.1" />
+    <PackageReference Include="SocketIoClientDotNet" Version="0.9.13" />
   </ItemGroup>
   <ItemGroup>
     <None Update="appsettings.json">
